requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
discord.py==2.2.2
dlib==19.24.1
docopt==0.6.2
face-recognition==1.3.0
face-recognition-models==0.3.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.3
Pillow==9.5.0
pipreqs==0.4.13
python-dotenv==1.0.0
requests==2.29.0
urllib3==1.26.15
yarg==0.1.9
yarl==1.9.2
